--- src/main/kotlin/com/server/glol/domain/match/dto/MatchDto.kt
@@ -1,9 +1,0 @@
-package com.server.glol.domain.match.dto
-
-class MatchDto(
-        val metadata: MetadataDto,
-        val info: InfoDto
-) {
-        constructor(): this(MetadataDto(), InfoDto())
-
-}
