+++ src/main/kotlin/com/server/glol/domain/match/dto/riot/matchv5/MatchDto.kt
@@ -1,0 +1,9 @@
+package com.server.glol.domain.match.dto.riot.matchv5
+
+class MatchDto(
+    val metadata: MetadataDto,
+    val info: InfoDto
+) {
+        constructor(): this(MetadataDto(), InfoDto())
+
+}
