+++ src/main/kotlin/com/server/glol/domain/match/dto/riot/matchv5/MetadataDto.kt
@@ -1,4 +1,4 @@
-package com.server.glol.domain.match.dto
+package com.server.glol.domain.match.dto.riot.matchv5
 
 class MetadataDto(
     val matchId: String
